+++ changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [x] fix: module marker exception usage
 - [x] enhancement: switch `merge` method from recursion to cycle
 - [x] enhancement: switch `flat` method from recursion to cycle
+- [x] refactoring: move `CaseInsensitiveDict` tests to other unit tests file
 
 ## [1.0.0] - 2023-01-27